rag-llm-microservice/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.35.0
torch==2.1.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
numpy==1.24.3
python-multipart==0.0.6
pytest==7.4.0
httpx==0.25.0